added customer search

+++ app/src/main/java/sk/scolopax/reservrant/data/Customer.java
@@ -14,26 +14,26 @@ public class Customer implements Parcelable {
     public static final String JSON_KEY_NAME_FIRST = "customerFirstName";
     public static final String JSON_KEY_NAME_LAST = "customerLastName";
 
-    public final int idCustomer;
+    public final long idCustomer;
     public final String nameFirst;
     public final String nameLast;
 
     /* Constructors */
 
-    public Customer(int idCustomer, String nameFirst, String nameLast) {
+    public Customer(long idCustomer, String nameFirst, String nameLast) {
         this.idCustomer = idCustomer;
         this.nameFirst = nameFirst;
         this.nameLast = nameLast;
     }
 
     public Customer(Cursor cursor) {
-        this.idCustomer = cursor.getInt(DatabaseContract.TableCustomers.COL_IDX_ID);
+        this.idCustomer = cursor.getLong(DatabaseContract.TableCustomers.COL_IDX_ID);
         this.nameFirst = cursor.getString(DatabaseContract.TableCustomers.COL_IDX_NAME_FIRST);
         this.nameLast = cursor.getString(DatabaseContract.TableCustomers.COL_IDX_NAME_LAST);
     }
 
     protected Customer(Parcel in) {
-        this.idCustomer = in.readInt();
+        this.idCustomer = in.readLong();
         this.nameFirst = in.readString();
         this.nameLast = in.readString();
     }
@@ -47,7 +47,7 @@ public class Customer implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(idCustomer);
+        dest.writeLong(idCustomer);
         dest.writeString(nameFirst);
         dest.writeString(nameLast);
     }
